salario mask

+++ src/database/migrations/2020_01_06_170705_create_funcionario_table.php
@@ -23,7 +23,7 @@ class CreateFuncionarioTable extends Migration
             $table->string('cpf');
             $table->string('password');
             $table->string('cargo',150);
-            $table->integer('salario');
+            $table->bigInteger('salario');
             $table->boolean('situacao');
             $table->rememberToken();
             $table->timestamps();
